Inspect long/wide/imputed data

Inspect long/wide/imputed data

inspectData(
  data,
  exposure,
  exposure_time_pts,
  outcome,
  ti_confounders,
  tv_confounders,
  epochs = NULL,
  hi_lo_cut = NULL,
  reference = NULL,
  comparison = NULL,
  home_dir = NA,
  verbose = TRUE,
  save.out = TRUE
)

Arguments

data

data in wide format as: a data frame, list of imputed data frames, or mids object

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

ti_confounders

list of time invariant confounders

tv_confounders

list of time-varying confounders with ".timepoint" suffix (time-varying exposure variables requires to be listed here)

epochs

(optional) data frame of exposure epoch labels and values

hi_lo_cut

(optional) list of two numbers indicating quantile values that reflect high and low values, respectively, for continuous exposure (default is median split)

reference

(optional) list of one or more strings of "-"-separated "l" and "h" values indicative of a reference exposure history to which to compare comparison, required if comparison is specified

comparison

(optional) list of one or more strings of "-"-separated "l" and "h" values indicative of comparison history/histories to compare to reference, required if reference is specified

home_dir

(optional) path to home directory (required if save.out = TRUE)

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

none

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3"),
            ti_confounders = "C",
            save.out = FALSE)
#> Downloading GitHub repo istallworthy/devMSMs@HEAD
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/private/var/folders/9h/vszz760d3v98y8ymym8rtwj00000gp/T/RtmpxI6pHy/remotesb0129f8ef89/istallworthy-devMSMs-0f94d34/DESCRIPTION’ ... OK
#> * preparing ‘devMSMs’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘devMSMs_0.0.0.9000.tar.gz’
#> 
#> Installing package into ‘/private/var/folders/9h/vszz760d3v98y8ymym8rtwj00000gp/T/RtmpsbSVoM/temp_libpathad345e789266’
#> (as ‘lib’ is unspecified)
#> USER ALERT: Below are the 4 variables spanning 2 unique domains that will be treated as confounding variables for the relation between A and D.3. 
#> Please inspect this list carefully. It should include all time-varying covariates, time invariant covariates, as well as lagged levels of exposure and outcome variables if they were collected at time points earlier than the outcome time point.
#> [1] "A.1" "A.2" "A.3" "C"  
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 48 (96%) individuals that fall into 8 out of the 8 total user-defined exposure histories created from median split values for low and high levels of exposure A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is sufficient spread to avoid extrapolation and low precision: 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3 containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h-h   |  8|
#> |h-h-l   |  6|
#> |h-l-h   |  4|
#> |h-l-l   |  7|
#> |l-h-h   |  4|
#> |l-h-l   |  6|
#> |l-l-h   |  7|
#> |l-l-l   |  6|
#> 
#> Your outcome variable(s) have the following type(s): data.frame
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            save.out = FALSE)
#> Skipping install of 'devMSMs' from a github remote, the SHA1 (0f94d34d) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> USER ALERT: Below are the 7 variables spanning 3 unique domains that will be treated as confounding variables for the relation between A and D.3. 
#> Please inspect this list carefully. It should include all time-varying covariates, time invariant covariates, as well as lagged levels of exposure and outcome variables if they were collected at time points earlier than the outcome time point.
#> [1] "A.1" "A.2" "A.3" "B.1" "B.2" "B.3" "C"  
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 48 (96%) individuals that fall into 8 out of the 8 total user-defined exposure histories created from median split values for low and high levels of exposure A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is sufficient spread to avoid extrapolation and low precision: 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3 containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h-h   |  8|
#> |h-h-l   |  6|
#> |h-l-h   |  4|
#> |h-l-l   |  7|
#> |l-h-h   |  4|
#> |l-h-l   |  6|
#> |l-l-h   |  7|
#> |l-l-l   |  6|
#> 
#> Your outcome variable(s) have the following type(s): data.frame
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            hi_lo_cut = c(0.8, 0.2),
            save.out = FALSE)
#> Skipping install of 'devMSMs' from a github remote, the SHA1 (0f94d34d) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> USER ALERT: Below are the 7 variables spanning 3 unique domains that will be treated as confounding variables for the relation between A and D.3. 
#> Please inspect this list carefully. It should include all time-varying covariates, time invariant covariates, as well as lagged levels of exposure and outcome variables if they were collected at time points earlier than the outcome time point.
#> [1] "A.1" "A.2" "A.3" "B.1" "B.2" "B.3" "C"  
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 3 (6%) individuals that fall into 2 out of the 8 the total user-defined exposure histories created from 20th and 80th percentile values for low and high levels of exposure A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is sufficient spread to avoid extrapolation and low precision: 
#> Warning: USER ALERT: There are no individuals in your sample that fall into h-h-h & h-h-l & h-l-h & l-h-h & l-h-l & l-l-h exposure history/histories. You may wish to consider different high/low cutoffs (for continuous exposures), alternative epochs, or choose a different measure to avoid extrapolation.
#> 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3 containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-l-l   |  1|
#> |l-l-l   |  2|
#> 
#> Your outcome variable(s) have the following type(s): data.frame
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            hi_lo_cut = c(0.8, 0.2),
            reference = "l-l-l",
            comparison = "h-h-h",
            save.out = FALSE)
#> Skipping install of 'devMSMs' from a github remote, the SHA1 (0f94d34d) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> USER ALERT: Below are the 7 variables spanning 3 unique domains that will be treated as confounding variables for the relation between A and D.3. 
#> Please inspect this list carefully. It should include all time-varying covariates, time invariant covariates, as well as lagged levels of exposure and outcome variables if they were collected at time points earlier than the outcome time point.
#> [1] "A.1" "A.2" "A.3" "B.1" "B.2" "B.3" "C"  
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 2 (4%) individuals that fall into 1 out of the 2 the total user-defined exposure histories created from 20th and 80th percentile values for low and high levels of exposure A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is sufficient spread to avoid extrapolation and low precision: 
#> Warning: USER ALERT: There are no individuals in your sample that fall into h-h-h exposure history/histories. You may wish to consider different high/low cutoffs (for continuous exposures), alternative epochs, or choose a different measure to avoid extrapolation.
#> 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3 containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |l-l-l   |  2|
#> 
#> Your outcome variable(s) have the following type(s): data.frame
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                                values = I(list(c(1, 2), c(3)))),
            save.out = FALSE)
#> Skipping install of 'devMSMs' from a github remote, the SHA1 (0f94d34d) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> USER ALERT: Below are the 7 variables spanning 3 unique domains that will be treated as confounding variables for the relation between A and D.3. 
#> Please inspect this list carefully. It should include all time-varying covariates, time invariant covariates, as well as lagged levels of exposure and outcome variables if they were collected at time points earlier than the outcome time point.
#> [1] "A.1" "A.2" "A.3" "B.1" "B.2" "B.3" "C"  
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 50 (100%) individuals that fall into 4 out of the 4 total user-defined exposure histories created from median split values for low and high levels of exposure A, respectively, across Infancy, Toddlerhood. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is sufficient spread to avoid extrapolation and low precision: 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects Infancy, Toddlerhood containing time points c(1, 2), 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h     | 10|
#> |h-l     | 15|
#> |l-h     | 15|
#> |l-l     | 10|
#> 
#> Your outcome variable(s) have the following type(s): data.frame